[기능 수정] 크루 리스트 컴포넌트 세분화
